+++ CHANGELOG.md
@@ -8,10 +8,14 @@ LogicNG uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Convenience methods `isSatisfiable`, `implies`, `isImpliedBy` and `isEquivalentTo` in the `Formula` class.
 - Two overloaded factory methods `mk` in `MiniSat` to construct a solver by formula factory, solver style and optional configuration.
+- Methods to directly apply Boolean functions on BDDs
+- Added `toFormula` method on BDDs to generate a formula via Shannon expansion
 
 ### Changed
 
-- Methods `generateFromCnf(Formula formula)` and `generateFromCnf(Collection<Formula> formulas)` in `ConstraintGraphGenerator` are now deprecated, since the constraint graph generation is not CNF specific. Both methods will be removed with LogicNG 3.0. Instead, use the general method `generateFromFormulas(Collection<Formula> formulas)`.
+- Methods `generateFromCnf(Formula formula)` and `generateFromCnf(Collection<Formula> formulas)` in `ConstraintGraphGenerator` are now deprecated, since the
+  constraint graph generation is not CNF specific. Both methods will be removed with LogicNG 3.0. Instead, use the general
+  method `generateFromFormulas(Collection<Formula> formulas)`.
 
 ## [2.3.2] - 2022-08-02